sound almost ok

+++ Kernel/include/sound.h
@@ -1,7 +1,7 @@
 #ifndef SOUND_H
 #define SOUND_H
 
-void beep(uint32_t freq, int time);
+void beep(uint64_t freq, uint64_t time);
 
 #include <time.h>
 #include <lib.h>
