+++ Userland/ShellCodeModule/include/library.h
@@ -34,4 +34,4 @@ int64_t puts(const char *s);
 
 char *gets(char *buf, uint16_t max_len);
 
-int64_t beep(uint32_t frequency, int time);
+int64_t beep(uint64_t frequency, uint64_t time);
